+++ kubel.el
@@ -289,35 +289,7 @@ CMD is the command string to run."
 (defvar kubel-selector-history '()
   "List of previously used selectors.")
 
-;; fallback list of resources if the version of kubectl doesn't support api-resources command
-(defvar kubel-kubernetes-resources-list
-  '("Pods"
-    "Services"
-    "Namespaces"
-    "Nodes"
-    "Configmaps"
-    "Secrets"
-    "Bindings"
-    "PersistentVolumeClaims"
-    "PersistentVolumes"
-    "ReplicationControllers"
-    "ResourceQuotas"
-    "ServiceAccounts"
-    "Deployments"
-    "DaemonSets"
-    "ReplicaSets"
-    "StatefulSets"
-    "Jobs"
-    "hpa"
-    "Images"
-    "Ingresses"
-    "ClusterRoles"
-    "RoleBindings"
-    "Roles"))
-
 (defvar-local kubel--kubernetes-version-cached nil)
-
-(defvar kubel--kubernetes-resources-list-cached nil)
 
 (defvar-local kubel--can-get-namespace-cached nil)
 
@@ -327,41 +299,19 @@ CMD is the command string to run."
 
 (defvar-local kubel--selected-items '())
 
+(defun kubel--kubernetes-resources-list ()
+  "Get list of resources from cache or from fetching the api resource."
+  (if (null kubel--kubernetes-resources-list-cached)
+      (setq kubel--kubernetes-resources-list-cached
+            (kubel--fetch-api-resource-list))
+    kubel--kubernetes-resources-list-cached))
+
 (defun kubel--invalidate-context-caches ()
   "Invalidate the context caches."
   (setq kubel--kubernetes-resources-list-cached nil)
-  (setq kubel--kubernetes-version-cached nil)
   (setq kubel--can-get-namespace-cached nil)
   (setq kubel--namespace-list-cached nil)
   (setq kubel--label-values-cached nil))
-
-(defun kubel-kubernetes-version ()
-  "Return a list with (major-version minor-version patch)."
-  (let* ((version-json (json-read-from-string
-                        (if (null kubel--kubernetes-version-cached)
-                            (setq kubel--kubernetes-version-cached
-                                  (kubel--exec-to-string (concat kubel-kubectl " version --output=json")))
-                           kubel--kubernetes-version-cached)))
-         (version-string (assoc-default 'gitVersion
-                                         (assoc-default 'serverVersion version-json))))
-    (string-match "v\\([0-9]*\\)\.\\([0-9]*\\)\.\\([0-9]*\\)[^0-9].*" version-string)
-    (list
-     (string-to-number (match-string 1 version-string))
-     (string-to-number (match-string 2 version-string))
-     (string-to-number (match-string 3 version-string)))))
-
-(defun kubel-kubernetes-compatible-p (version)
-  "Return TRUE if kubernetes version is greater than or equal to VERSION.
-VERSION should be a list of (major-version minor-version patch)."
-  (let*
-      ((kubernetes-version (kubel-kubernetes-version))
-       (kubernetes-major-version (nth 0 kubernetes-version))
-       (kubernetes-minor-version (nth 1 kubernetes-version))
-       (kubernetes-patch-version (nth 2 kubernetes-version)))
-    (and
-     (<= (nth 0 version) kubernetes-major-version)
-     (or (<= (nth 1 version) kubernetes-minor-version) (< (nth 0 version) kubernetes-major-version))
-     (or (<= (nth 2 version) kubernetes-patch-version) (< (nth 1 version) kubernetes-minor-version)))))
 
 (defun kubel--populate-list ()
   "Return a list with a tabulated list format and \"tabulated-list-entries\"."
@@ -911,12 +861,7 @@ the context caches, including the cached resource list."
   (interactive "P")
   (when refresh (kubel--invalidate-context-caches))
   (let* ((current-buffer-name (kubel--buffer-name))
-         (resource-list (if (kubel-kubernetes-compatible-p '(1 13 3))
-                            (if (null kubel--kubernetes-resources-list-cached)
-                                (setq kubel--kubernetes-resources-list-cached
-                                      (kubel--fetch-api-resource-list))
-                              kubel--kubernetes-resources-list-cached)
-                          kubel-kubernetes-resources-list))
+         (resource-list (kubel--kubernetes-resources-list))
          (kubel--buffer (get-buffer current-buffer-name))
          (last-default-directory (when kubel--buffer (with-current-buffer kubel--buffer default-directory))))
     (with-current-buffer (clone-buffer)
@@ -1137,7 +1082,7 @@ RESET is to be called if the search is nil after the first attempt."
   "Quickly edit any resource."
   (interactive)
   (kubel--describe-resource
-   (completing-read "Select resource: " kubel-kubernetes-resources-list)))
+   (completing-read "Select resource: " (kubel--kubernetes-resources-list))))
 
 (defun kubel-show-process-buffer ()
   "Show the kubel-process-buffer."